+++ screens/TbM/TbmStepForm/Step2.js
@@ -1,0 +1,277 @@
+import {
+  View,
+  Text,
+  Button,
+  TextInput,
+  ScrollView,
+  TouchableOpacity,
+  Alert,
+} from "react-native";
+import React, { useState, useEffect, useLayoutEffect } from "react";
+import { useNavigation, useRoute } from "@react-navigation/native";
+import { AntDesign, Entypo, FontAwesome } from "@expo/vector-icons";
+
+const Step2 = ({ onNext, onPrev }) => {
+  return (
+    <ScrollView
+      style={{
+        width: "100%",
+        height: "100%",
+        // justifyContent: "center",
+        flexDirection: "column",
+        // alignItems: "center",
+        backgroundColor: "white",
+      }}
+    >
+      <View
+        style={{
+          width: "100%",
+          height: 120,
+          backgroundColor: "#ffaa00a1",
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          alignItems: "center",
+        }}
+      >
+        <View>
+          <Text
+            style={{
+              color: "#2e2d6e",
+              fontSize: 22,
+              marginLeft: 0,
+              fontWeight: "600",
+            }}
+          >
+            2.) Tool-Box Talk
+          </Text>
+        </View>
+      </View>
+
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          Company Supervisor/Line Manager Name
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Please Enter Company Supervisor/Manager"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          Safety Representative
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Enter Safety Representative"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          Department
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Enter Department Name"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          Contractor Representative
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Enter Contractor Representative Name"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          Contractor Employees
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Contractor Employees Name"
+        />
+      </View>
+      <View
+        style={{
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          marginTop: 40,
+        }}
+      >
+        <TouchableOpacity
+          onPress={onPrev}
+          style={{
+            backgroundColor: "#2e2d6e",
+            padding: 10,
+            borderRadius: 50,
+          }}
+        >
+          {/* <Button title="Next" onPress={onNext} /> */}
+          <Text
+            style={{
+              paddingHorizontal: 20,
+              paddingVertical: 0,
+              fontSize: 16,
+              textAlign: "center",
+              fontWeight: "500",
+              color: "white",
+            }}
+          >
+            Prev
+          </Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={{
+            backgroundColor: "#2e2d6e",
+            padding: 10,
+            borderRadius: 50,
+            marginLeft: 10,
+          }}
+          onPress={onNext}
+        >
+          <View
+            style={{
+              alignItems: "center",
+              flexDirection: "row",
+              justifyContent: "center",
+            }}
+          >
+            <Text
+              style={{
+                paddingHorizontal: 20,
+                paddingVertical: 0,
+                fontSize: 16,
+                textAlign: "center",
+                fontWeight: "500",
+                color: "white",
+              }}
+            >
+              Next
+            </Text>
+            <AntDesign name="right" size={18} color="white" />
+          </View>
+        </TouchableOpacity>
+      </View>
+    </ScrollView>
+  );
+};
+
+export default Step2;
