+++ screens/TbM/TbmStepForm/Step4.js
@@ -1,0 +1,269 @@
+import {
+  View,
+  Text,
+  Button,
+  TextInput,
+  ScrollView,
+  TouchableOpacity,
+  Alert,
+} from "react-native";
+import React, { useState, useEffect, useLayoutEffect } from "react";
+import { useNavigation, useRoute } from "@react-navigation/native";
+import { AntDesign, Entypo, FontAwesome } from "@expo/vector-icons";
+
+const Step4 = ({ onNext, onPrev }) => {
+  return (
+    <ScrollView
+      style={{
+        width: "100%",
+        height: "100%",
+        // justifyContent: "center",
+        flexDirection: "column",
+        // alignItems: "center",
+        backgroundColor: "white",
+      }}
+    >
+      <View
+        style={{
+          width: "100%",
+          height: 120,
+          backgroundColor: "#ffaa00a1",
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          alignItems: "center",
+        }}
+      >
+        <View>
+          <Text
+            style={{
+              color: "#2e2d6e",
+              fontSize: 22,
+              marginLeft: 0,
+              fontWeight: "600",
+            }}
+          >
+            4.) Final Details
+          </Text>
+        </View>
+      </View>
+
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          4.) SOP - Relative to the group
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Type Your Message..."
+          multiline={true} // Allow multiline input
+          numberOfLines={2} // Set maximum number of lines
+          textAlignVertical="top"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          5.) Reminder to there Employees of there Personal Responsibilities to
+          insure and maintains.
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Type Your Message..."
+          multiline={true} // Allow multiline input
+          numberOfLines={2} // Set maximum number of lines
+          textAlignVertical="top"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+          }}
+        >
+          6.) Safety Message Hand-out/circular to be shared within contract
+          Employees.
+        </Text>
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Type Your Message..."
+          multiline={true} // Allow multiline input
+          numberOfLines={2} // Set maximum number of lines
+          textAlignVertical="top"
+        />
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 14,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+            paddingRight: 15,
+          }}
+        >
+          7.) Action resulting from meeting and points raised bye the contract
+          employees and supervisor.
+        </Text>
+      </View>
+      <View
+        style={{
+          marginTop: 20,
+          marginHorizontal: 20,
+          width: "100%",
+        }}
+      >
+        <TextInput
+          style={{
+            paddingHorizontal: 12,
+            paddingVertical: 12,
+            width: "90%",
+            backgroundColor: "#F5F5F5",
+            elevation: 3,
+            borderRadius: 5,
+            color: "black",
+          }}
+          placeholder="Type Your Message..."
+          multiline={true} // Allow multiline input
+          numberOfLines={2} // Set maximum number of lines
+          textAlignVertical="top"
+        />
+      </View>
+      <View
+        style={{
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          marginTop: 40,
+        }}
+      >
+        <TouchableOpacity
+          onPress={onPrev}
+          style={{
+            backgroundColor: "#2e2d6e",
+            padding: 10,
+            borderRadius: 50,
+          }}
+        >
+          {/* <Button title="Next" onPress={onNext} /> */}
+          <Text
+            style={{
+              paddingHorizontal: 20,
+              paddingVertical: 0,
+              fontSize: 16,
+              textAlign: "center",
+              fontWeight: "500",
+              color: "white",
+            }}
+          >
+            Prev
+          </Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={{
+            backgroundColor: "#2e2d6e",
+            padding: 10,
+            borderRadius: 50,
+            marginLeft: 10,
+          }}
+          onPress={onNext}
+        >
+          <View
+            style={{
+              alignItems: "center",
+              flexDirection: "row",
+              justifyContent: "center",
+            }}
+          >
+            <Text
+              style={{
+                paddingHorizontal: 20,
+                paddingVertical: 0,
+                fontSize: 16,
+                textAlign: "center",
+                fontWeight: "500",
+                color: "white",
+              }}
+            >
+              Next
+            </Text>
+            <AntDesign name="right" size={18} color="white" />
+          </View>
+        </TouchableOpacity>
+      </View>
+    </ScrollView>
+  );
+};
+
+export default Step4;
