+++ screens/TbM/TbmStepForm/Step5.js
@@ -1,0 +1,229 @@
+import {
+  View,
+  Text,
+  Button,
+  TextInput,
+  ScrollView,
+  TouchableOpacity,
+  Alert,
+} from "react-native";
+import React, { useState, useEffect, useLayoutEffect } from "react";
+import { useNavigation, useRoute } from "@react-navigation/native";
+import { MaterialIcons } from "@expo/vector-icons";
+
+const Step5 = ({ onNext, onPrev }) => {
+  const [inputList, setInputList] = useState([{ id: 1, text: "" }]);
+  const [idCounter, setIdCounter] = useState(2); // Counter for generating unique ids
+
+  const handleAddInput = () => {
+    const newInput = { id: idCounter, text: "" };
+    setInputList([...inputList, newInput]);
+    setIdCounter(idCounter + 1);
+  };
+
+  const handleRemoveInput = (idToRemove) => {
+    const updatedList = inputList.filter((item) => item.id !== idToRemove);
+    setInputList(updatedList);
+  };
+
+  const handleInputChange = (text, id) => {
+    const updatedList = inputList.map((item) =>
+      item.id === id ? { ...item, text: text } : item
+    );
+    setInputList(updatedList);
+  };
+  return (
+    <ScrollView
+      style={{
+        width: "100%",
+        height: "100%",
+        // justifyContent: "center",
+        flexDirection: "column",
+        // alignItems: "center",
+        backgroundColor: "white",
+      }}
+    >
+      <View
+        style={{
+          width: "100%",
+          height: 120,
+          backgroundColor: "#ffaa00a1",
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          alignItems: "center",
+        }}
+      >
+        <View>
+          <Text
+            style={{
+              color: "#2e2d6e",
+              fontSize: 22,
+              marginLeft: 0,
+              fontWeight: "600",
+            }}
+          >
+            5.) TBT Attendance
+          </Text>
+        </View>
+      </View>
+
+      <View
+        style={{
+          flex: 1,
+          //   justifyContent: "center",
+          //   alignItems: "center",
+          padding: 20,
+        }}
+      >
+        <Text
+          style={{
+            textAlign: "left",
+            fontSize: 16,
+            paddingHorizontal: 5,
+            paddingVertical: 5,
+            fontWeight: "600",
+            color: "#00308F",
+            paddingRight: 15,
+            marginBottom: 10,
+          }}
+        >
+          Enter Employee Names
+        </Text>
+
+        {inputList.map((input, index) => (
+          <View
+            key={input.id}
+            style={{
+              flexDirection: "row",
+              alignItems: "center",
+              marginBottom: 10,
+            }}
+          >
+            <TextInput
+              style={{
+                paddingHorizontal: 12,
+                paddingVertical: 12,
+                width: "90%",
+                backgroundColor: "#F5F5F5",
+                elevation: 3,
+                borderRadius: 5,
+                color: "black",
+              }}
+              value={input.text}
+              onChangeText={(text) => handleInputChange(text, input.id)}
+              placeholder={`Employee Name ${index + 1}`}
+            />
+            {index > 0 && ( // Render remove button for all inputs except the first one
+              <TouchableOpacity
+                onPress={() => handleRemoveInput(input.id)}
+                style={{
+                  marginTop: 10,
+                  backgroundColor: "#244aca",
+                  paddingHorizontal: 20,
+                  paddingVertical: 10,
+                  borderRadius: 50,
+                  marginLeft: 10,
+                }}
+              >
+                <Text
+                  style={{
+                    fontSize: 14,
+                    fontWeight: "600",
+                    color: "#fff",
+                  }}
+                >
+                  -
+                </Text>
+              </TouchableOpacity>
+            )}
+          </View>
+        ))}
+        {/* <Button title="+" onPress={handleAddInput} /> */}
+        <TouchableOpacity
+          onPress={handleAddInput}
+          style={{
+            marginTop: 10,
+            backgroundColor: "#244aca",
+            paddingHorizontal: 20,
+            paddingVertical: 10,
+            borderRadius: 50,
+          }}
+        >
+          <Text
+            style={{
+              fontSize: 14,
+              fontWeight: "600",
+              color: "#fff",
+            }}
+          >
+            + Add Employee
+          </Text>
+        </TouchableOpacity>
+      </View>
+      <View
+        style={{
+          flexDirection: "row",
+          justifyContent: "space-between",
+          paddingHorizontal: 20,
+          marginTop: 40,
+        }}
+      >
+        <TouchableOpacity
+          onPress={onPrev}
+          style={{
+            backgroundColor: "#2e2d6e",
+            padding: 10,
+            borderRadius: 50,
+          }}
+        >
+          {/* <Button title="Next" onPress={onNext} /> */}
+          <Text
+            style={{
+              paddingHorizontal: 20,
+              paddingVertical: 0,
+              fontSize: 16,
+              textAlign: "center",
+              fontWeight: "500",
+              color: "white",
+            }}
+          >
+            Prev
+          </Text>
+        </TouchableOpacity>
+        <TouchableOpacity
+          style={{
+            backgroundColor: "green",
+            padding: 10,
+            borderRadius: 50,
+            marginLeft: 10,
+          }}
+          onPress={onNext}
+        >
+          <View
+            style={{
+              alignItems: "center",
+              flexDirection: "row",
+              justifyContent: "center",
+            }}
+          >
+            <Text
+              style={{
+                paddingHorizontal: 20,
+                paddingVertical: 0,
+                fontSize: 16,
+                textAlign: "center",
+                fontWeight: "500",
+                color: "white",
+              }}
+            >
+              SUBMIT
+            </Text>
+            <MaterialIcons name="done" size={18} color="white" />
+          </View>
+        </TouchableOpacity>
+      </View>
+    </ScrollView>
+  );
+};
+export default Step5;
