+++ screens/TbM/TbtForm.js
@@ -1,1318 +1,11 @@
-import {
-  View,
-  Text,
-  Button,
-  TextInput,
-  ScrollView,
-  TouchableOpacity,
-  Alert,
-} from "react-native";
-import React, { useState, useEffect, useLayoutEffect } from "react";
-import { useNavigation, useRoute } from "@react-navigation/native";
-import { AntDesign, Entypo, FontAwesome } from "@expo/vector-icons";
+import { View } from "react-native";
+import React, { useState } from "react";
 
-const Step1 = ({ onNext }) => {
-  const currentDate = new Date();
-
-  const [permitNumber, setPermitNumber] = useState("");
-
-  const [currentTime, setCurrentTime] = useState(new Date());
-
-  const [shift, setShift] = useState("");
-
-  useEffect(() => {
-    // Update the current time every second
-    const timerID = setInterval(() => {
-      setCurrentTime(new Date());
-    }, 1000);
-
-    // Clean up the timer
-    return () => {
-      clearInterval(timerID);
-    };
-  }, []);
-  // Extract hours, minutes, and seconds
-  const hours = currentTime.getHours();
-  const minutes = currentTime.getMinutes();
-  const seconds = currentTime.getSeconds();
-
-  useEffect(() => {
-    if (hours > 6 && hours < 8) {
-      setShift("Morning A Shift");
-    } else if (hours > 8 && hours < 12) {
-      setShift("Morning General Shift");
-    } else if (hours > 12 && hours < 17) {
-      setShift("Afternoon B Shift");
-    } else {
-      setShift("Night C Shift");
-    }
-  }, []);
-
-  const navigation = useNavigation();
-
-  useEffect(() => {
-    navigation.getParent()?.setOptions({
-      tabBarStyle: {
-        display: "none",
-      },
-    });
-  }, [navigation]);
-
-  // Extract day, month (in number), and year
-  const day = currentDate.getDate(); // Returns the day of the month (1-31)
-  const month = currentDate.getMonth() + 1; // Returns the month (0-11); adding 1 to get the actual month number
-  const year = currentDate.getFullYear(); // Returns the year (e.g., 2024)
-
-  return (
-    <ScrollView
-      style={{
-        width: "100%",
-        height: "100%",
-        // justifyContent: "center",
-        flexDirection: "column",
-        // alignItems: "center",
-        backgroundColor: "white",
-      }}
-    >
-      <View
-        style={{
-          width: "100%",
-          height: 120,
-          backgroundColor: "#ffaa00a1",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          alignItems: "center",
-        }}
-      >
-        <TouchableOpacity
-          onPress={() => {
-            navigation.navigate("ToolBoxTalk");
-          }}
-        >
-          <AntDesign name="left" size={25} color="#2e2d6e" />
-        </TouchableOpacity>
-        <View>
-          <Text
-            style={{
-              color: "#2e2d6e",
-              fontSize: 22,
-              marginLeft: 20,
-              fontWeight: "600",
-            }}
-          >
-            Tool Box Meeting
-          </Text>
-        </View>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Today's Date
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          value={`${day}/${month}/${year}`}
-          placeholder="Date"
-          editable={false}
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Current Time
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Time"
-          value={`${hours}:${minutes < 10 ? `0${minutes}` : minutes}:${
-            seconds < 10 ? `0${seconds}` : seconds
-          }`}
-          editable={false}
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Current Shift
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Shift"
-          value={shift}
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Site Location
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Enter Your Site Location"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Work Order / Permit Number
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          value={permitNumber}
-          placeholder="Enter Your Permit Number"
-          onChangeText={(permitNumber) => {
-            setPermitNumber(permitNumber);
-          }}
-          keyboardType="numeric"
-        />
-      </View>
-      <View
-        style={{
-          flex: 1,
-          justifyContent: "center",
-          alignItems: "center",
-          marginTop: 80,
-        }}
-      >
-        <TouchableOpacity
-          onPress={() => {
-            if (permitNumber.trim() === "") {
-              // Show alert message if text is empty
-              Alert.alert("Error", "Please Enter a Permit Number");
-            } else {
-              // Show success message if text is not empty
-              onNext();
-            }
-          }}
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-            position: "absolute",
-            bottom: 0,
-            right: 20,
-          }}
-        >
-          {/* <Button title="Next" onPress={onNext} /> */}
-          <View
-            style={{
-              alignItems: "center",
-              flexDirection: "row",
-              justifyContent: "center",
-            }}
-          >
-            <Text
-              style={{
-                paddingHorizontal: 20,
-                paddingVertical: 0,
-                fontSize: 16,
-                textAlign: "center",
-                fontWeight: "500",
-                color: "white",
-              }}
-            >
-              Next
-            </Text>
-            <AntDesign name="right" size={18} color="white" />
-          </View>
-        </TouchableOpacity>
-      </View>
-    </ScrollView>
-  );
-};
-
-const Step2 = ({ onNext, onPrev }) => {
-  return (
-    <ScrollView
-      style={{
-        width: "100%",
-        height: "100%",
-        // justifyContent: "center",
-        flexDirection: "column",
-        // alignItems: "center",
-        backgroundColor: "white",
-      }}
-    >
-      <View
-        style={{
-          width: "100%",
-          height: 120,
-          backgroundColor: "#ffaa00a1",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          alignItems: "center",
-        }}
-      >
-        <View>
-          <Text
-            style={{
-              color: "#2e2d6e",
-              fontSize: 22,
-              marginLeft: 0,
-              fontWeight: "600",
-            }}
-          >
-            2.) Tool-Box Talk
-          </Text>
-        </View>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Company Supervisor/Line Manager Name
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Please Enter Company Supervisor/Manager"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Safety Representative
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Enter Safety Representative"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Department
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Enter Department Name"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Contractor Representative
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Enter Contractor Representative Name"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          Contractor Employees
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Contractor Employees Name"
-        />
-      </View>
-      <View
-        style={{
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          marginTop: 40,
-        }}
-      >
-        <TouchableOpacity
-          onPress={onPrev}
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-          }}
-        >
-          {/* <Button title="Next" onPress={onNext} /> */}
-          <Text
-            style={{
-              paddingHorizontal: 20,
-              paddingVertical: 0,
-              fontSize: 16,
-              textAlign: "center",
-              fontWeight: "500",
-              color: "white",
-            }}
-          >
-            Prev
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-            marginLeft: 10,
-          }}
-          onPress={onNext}
-        >
-          <View
-            style={{
-              alignItems: "center",
-              flexDirection: "row",
-              justifyContent: "center",
-            }}
-          >
-            <Text
-              style={{
-                paddingHorizontal: 20,
-                paddingVertical: 0,
-                fontSize: 16,
-                textAlign: "center",
-                fontWeight: "500",
-                color: "white",
-              }}
-            >
-              Next
-            </Text>
-            <AntDesign name="right" size={18} color="white" />
-          </View>
-        </TouchableOpacity>
-      </View>
-    </ScrollView>
-  );
-};
-
-const Step3 = ({ onNext, onPrev }) => {
-  return (
-    <ScrollView
-      style={{
-        width: "100%",
-        height: "100%",
-        // justifyContent: "center",
-        flexDirection: "column",
-        // alignItems: "center",
-        backgroundColor: "white",
-      }}
-    >
-      <View
-        style={{
-          width: "100%",
-          height: 120,
-          backgroundColor: "#ffaa00a1",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          alignItems: "center",
-        }}
-      >
-        <View>
-          <Text
-            style={{
-              color: "#2e2d6e",
-              fontSize: 22,
-              marginLeft: 0,
-              fontWeight: "600",
-            }}
-          >
-            3.) Safety Contract & Reviews
-          </Text>
-        </View>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          1.) Safety Contract review of action items from last meetings.
-        </Text>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type your answer..."
-          multiline={true} // Allow multiline input
-          numberOfLines={4} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 10,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          2.) Items of General Safety Importance to the total work site.
-        </Text>
-        <Text
-          style={{
-            marginHorizontal: 10,
-            marginRight: 50,
-            fontSize: 12,
-            marginBottom: 20,
-            color: "gray",
-            fontWeight: "300",
-            color: "#2c8fff",
-          }}
-        >
-          Ask employee to mention any Incident/near miss duration that past day
-          which may have or have resulted into damage to property or injury yo
-          company or Contract personnel
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type your answer..."
-          multiline={true} // Allow multiline input
-          numberOfLines={4} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          3.) Items of Safety Interest to this Group.
-        </Text>
-        <Text
-          style={{
-            marginHorizontal: 10,
-            marginRight: 50,
-            fontSize: 12,
-            marginBottom: 0,
-            color: "gray",
-            fontWeight: "300",
-            color: "#2c8fff",
-          }}
-        >
-          Red Strips, Orange Strips, Green Strips, Safety-alert tips for Satefy
-          communications, hazards or Safety conditions applicable to this
-          group's work area
-        </Text>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Please provide any queries you may have on the above..."
-          multiline={true} // Allow multiline input
-          numberOfLines={4} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          marginTop: 40,
-        }}
-      >
-        <TouchableOpacity
-          onPress={onPrev}
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-          }}
-        >
-          {/* <Button title="Next" onPress={onNext} /> */}
-          <Text
-            style={{
-              paddingHorizontal: 20,
-              paddingVertical: 0,
-              fontSize: 16,
-              textAlign: "center",
-              fontWeight: "500",
-              color: "white",
-            }}
-          >
-            Prev
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-            marginLeft: 10,
-          }}
-          onPress={onNext}
-        >
-          <View
-            style={{
-              alignItems: "center",
-              flexDirection: "row",
-              justifyContent: "center",
-            }}
-          >
-            <Text
-              style={{
-                paddingHorizontal: 20,
-                paddingVertical: 0,
-                fontSize: 16,
-                textAlign: "center",
-                fontWeight: "500",
-                color: "white",
-              }}
-            >
-              Next
-            </Text>
-            <AntDesign name="right" size={18} color="white" />
-          </View>
-        </TouchableOpacity>
-      </View>
-    </ScrollView>
-  );
-};
-
-const Step4 = ({ onNext, onPrev }) => {
-  return (
-    <ScrollView
-      style={{
-        width: "100%",
-        height: "100%",
-        // justifyContent: "center",
-        flexDirection: "column",
-        // alignItems: "center",
-        backgroundColor: "white",
-      }}
-    >
-      <View
-        style={{
-          width: "100%",
-          height: 120,
-          backgroundColor: "#ffaa00a1",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          alignItems: "center",
-        }}
-      >
-        <View>
-          <Text
-            style={{
-              color: "#2e2d6e",
-              fontSize: 22,
-              marginLeft: 0,
-              fontWeight: "600",
-            }}
-          >
-            4.) Final Details
-          </Text>
-        </View>
-      </View>
-
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          4.) SOP - Relative to the group
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type Your Message..."
-          multiline={true} // Allow multiline input
-          numberOfLines={2} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          5.) Reminder to there Employees of there Personal Responsibilities to
-          insure and maintains.
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type Your Message..."
-          multiline={true} // Allow multiline input
-          numberOfLines={2} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-          }}
-        >
-          6.) Safety Message Hand-out/circular to be shared within contract
-          Employees.
-        </Text>
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type Your Message..."
-          multiline={true} // Allow multiline input
-          numberOfLines={2} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 14,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-            paddingRight: 15,
-          }}
-        >
-          7.) Action resulting from meeting and points raised bye the contract
-          employees and supervisor.
-        </Text>
-      </View>
-      <View
-        style={{
-          marginTop: 20,
-          marginHorizontal: 20,
-          width: "100%",
-        }}
-      >
-        <TextInput
-          style={{
-            paddingHorizontal: 12,
-            paddingVertical: 12,
-            width: "90%",
-            backgroundColor: "#F5F5F5",
-            elevation: 3,
-            borderRadius: 5,
-            color: "black",
-          }}
-          placeholder="Type Your Message..."
-          multiline={true} // Allow multiline input
-          numberOfLines={2} // Set maximum number of lines
-          textAlignVertical="top"
-        />
-      </View>
-      <View
-        style={{
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          marginTop: 40,
-        }}
-      >
-        <TouchableOpacity
-          onPress={onPrev}
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-          }}
-        >
-          {/* <Button title="Next" onPress={onNext} /> */}
-          <Text
-            style={{
-              paddingHorizontal: 20,
-              paddingVertical: 0,
-              fontSize: 16,
-              textAlign: "center",
-              fontWeight: "500",
-              color: "white",
-            }}
-          >
-            Prev
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-            marginLeft: 10,
-          }}
-          onPress={onNext}
-        >
-          <View
-            style={{
-              alignItems: "center",
-              flexDirection: "row",
-              justifyContent: "center",
-            }}
-          >
-            <Text
-              style={{
-                paddingHorizontal: 20,
-                paddingVertical: 0,
-                fontSize: 16,
-                textAlign: "center",
-                fontWeight: "500",
-                color: "white",
-              }}
-            >
-              Next
-            </Text>
-            <AntDesign name="right" size={18} color="white" />
-          </View>
-        </TouchableOpacity>
-      </View>
-    </ScrollView>
-  );
-};
-
-const Step5 = ({ onNext, onPrev }) => {
-  const [inputList, setInputList] = useState([{ id: 1, text: "" }]);
-  const [idCounter, setIdCounter] = useState(2); // Counter for generating unique ids
-
-  const handleAddInput = () => {
-    const newInput = { id: idCounter, text: "" };
-    setInputList([...inputList, newInput]);
-    setIdCounter(idCounter + 1);
-  };
-
-  const handleRemoveInput = (idToRemove) => {
-    const updatedList = inputList.filter((item) => item.id !== idToRemove);
-    setInputList(updatedList);
-  };
-
-  const handleInputChange = (text, id) => {
-    const updatedList = inputList.map((item) =>
-      item.id === id ? { ...item, text: text } : item
-    );
-    setInputList(updatedList);
-  };
-  return (
-    <ScrollView
-      style={{
-        width: "100%",
-        height: "100%",
-        // justifyContent: "center",
-        flexDirection: "column",
-        // alignItems: "center",
-        backgroundColor: "white",
-      }}
-    >
-      <View
-        style={{
-          width: "100%",
-          height: 120,
-          backgroundColor: "#ffaa00a1",
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          alignItems: "center",
-        }}
-      >
-        <View>
-          <Text
-            style={{
-              color: "#2e2d6e",
-              fontSize: 22,
-              marginLeft: 0,
-              fontWeight: "600",
-            }}
-          >
-            5.) TBT Attendance
-          </Text>
-        </View>
-      </View>
-
-      <View
-        style={{
-          flex: 1,
-          //   justifyContent: "center",
-          //   alignItems: "center",
-          padding: 20,
-        }}
-      >
-        <Text
-          style={{
-            textAlign: "left",
-            fontSize: 16,
-            paddingHorizontal: 5,
-            paddingVertical: 5,
-            fontWeight: "600",
-            color: "#00308F",
-            paddingRight: 15,
-            marginBottom: 10,
-          }}
-        >
-          Enter Employee Names
-        </Text>
-
-        {inputList.map((input, index) => (
-          <View
-            key={input.id}
-            style={{
-              flexDirection: "row",
-              alignItems: "center",
-              marginBottom: 10,
-            }}
-          >
-            <TextInput
-              style={{
-                paddingHorizontal: 12,
-                paddingVertical: 12,
-                width: "90%",
-                backgroundColor: "#F5F5F5",
-                elevation: 3,
-                borderRadius: 5,
-                color: "black",
-              }}
-              value={input.text}
-              onChangeText={(text) => handleInputChange(text, input.id)}
-              placeholder={`Employee Name ${index + 1}`}
-            />
-            {index > 0 && ( // Render remove button for all inputs except the first one
-              <TouchableOpacity
-                onPress={() => handleRemoveInput(input.id)}
-                style={{
-                  marginTop: 10,
-                  backgroundColor: "#244aca",
-                  paddingHorizontal: 20,
-                  paddingVertical: 10,
-                  borderRadius: 50,
-                  marginLeft: 10,
-                }}
-              >
-                <Text
-                  style={{
-                    fontSize: 14,
-                    fontWeight: "600",
-                    color: "#fff",
-                  }}
-                >
-                  -
-                </Text>
-              </TouchableOpacity>
-            )}
-          </View>
-        ))}
-        {/* <Button title="+" onPress={handleAddInput} /> */}
-        <TouchableOpacity
-          onPress={handleAddInput}
-          style={{
-            marginTop: 10,
-            backgroundColor: "#244aca",
-            paddingHorizontal: 20,
-            paddingVertical: 10,
-            borderRadius: 50,
-          }}
-        >
-          <Text
-            style={{
-              fontSize: 14,
-              fontWeight: "600",
-              color: "#fff",
-            }}
-          >
-            + Add Employee
-          </Text>
-        </TouchableOpacity>
-      </View>
-      <View
-        style={{
-          flexDirection: "row",
-          justifyContent: "space-between",
-          paddingHorizontal: 20,
-          marginTop: 40,
-        }}
-      >
-        <TouchableOpacity
-          onPress={onPrev}
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-          }}
-        >
-          {/* <Button title="Next" onPress={onNext} /> */}
-          <Text
-            style={{
-              paddingHorizontal: 20,
-              paddingVertical: 0,
-              fontSize: 16,
-              textAlign: "center",
-              fontWeight: "500",
-              color: "white",
-            }}
-          >
-            Prev
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity
-          style={{
-            backgroundColor: "#2e2d6e",
-            padding: 10,
-            borderRadius: 50,
-            marginLeft: 10,
-          }}
-          onPress={onNext}
-        >
-          <Text
-            style={{
-              paddingHorizontal: 20,
-              paddingVertical: 0,
-              fontSize: 16,
-              textAlign: "center",
-              fontWeight: "500",
-              color: "white",
-            }}
-          >
-            Submit
-          </Text>
-        </TouchableOpacity>
-      </View>
-    </ScrollView>
-  );
-};
+import Step1 from "./TbmStepForm/Step1";
+import Step2 from "./TbmStepForm/Step2";
+import Step3 from "./TbmStepForm/Step3";
+import Step4 from "./TbmStepForm/Step4";
+import Step5 from "./TbmStepForm/Step5";
 
 const TbtForm = () => {
   const [step, setStep] = useState(1);
